replace `$PGUSER` and `$PGPASSWORD` with standard vars
The [upstream image](https://hub.docker.com/_/postgres/) uses
`$POSTGRES_USER` and `$POSTGRES_PASSWORD` for configuration; since the
scripts in this image use `$PGUSER` and `$PGPASSWORD`, it's laborious to
port other init scripts that were written for the upstream image over to
this image.  I've changed this image's scripts to bring them in line
with upstream, to make it easier to drop this image into place as a
replacement.